pyproject.toml
[tool.poetry]
name = "slac-conversations"
version = "0.1.0"
description = ""
authors = ["tanlda <ledinhanhtan.stack@gmail.com>"]
readme = "README.md"
packages = [{ include = "scripts", from = "." }]

[tool.poetry.scripts]
makemigrations = "scripts.django:makemigrations"
migrate = "scripts.django:migrate"
shell = "scripts.django:shell"
server = "scripts.django:runserver"
gunicorn = "scripts.django:gunicorn"


[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.7.2"
asttokens = "2.4.1"
certifi = "2024.2.2"
cfgv = "3.4.0"
charset-normalizer = "3.3.2"
decorator = "5.1.1"
distlib = "0.3.8"
django = "5.0.2"
django-cors-headers = "4.3.1"
django-environ = "0.11.2"
djangorestframework = "3.14.0"
executing = "2.0.1"
factory-boy = "3.3.0"
faker = "23.3.0"
filelock = "3.13.1"
gunicorn = "21.2.0"
identify = "2.5.35"
idna = "3.6"
iniconfig = "2.0.0"
ipython = "8.21.0"
jedi = "0.19.1"
matplotlib-inline = "0.1.6"
nodeenv = "1.8.0"
packaging = "23.2"
parso = "0.8.3"
pexpect = "4.9.0"
platformdirs = "4.2.0"
pluggy = "1.4.0"
pre-commit = "3.6.2"
prompt-toolkit = "3.0.43"
psycopg2 = "2.9.9"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pygments = "2.17.2"
pytest = "8.0.2"
pytest-django = "4.8.0"
pytest-env = "1.1.3"
pytest-mock = "3.12.0"
python-dateutil = "2.8.2"
pytz = "2024.1"
pyyaml = "6.0.1"
requests = "2.31.0"
shortuuid = "1.0.11"
six = "1.16.0"
split-settings = "1.0.0"
sqlparse = "0.4.4"
stack-data = "0.6.3"
traitlets = "5.14.1"
urllib3 = "2.2.1"
validators = "0.22.0"
virtualenv = "20.25.0"
wcwidth = "0.2.13"
microslac-core = "0.0.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120


[tool.pyright]
include = ["src"]
exclude = ["**/node_modules", "**/__pycache__"]
reportMissingImports = true
reportMissingTypeStubs = false
reportGeneralTypeIssues = false
reportOptionalSubscript = false
reportPrivateImportUsage = "information"
reportUnusedFunction = true

pythonPlatform = "Linux"
pythonVersion = "3.11"

[tool.isort]
known_local_folder = ["webmaster_verification", "sk*"]
known_typing_extensions = "typing_extensions"
line_length = 120
no_lines_before = "TYPING_EXTENSIONS"
profile = "black"
sections = [
    "FUTURE",
    "STDLIB",
    "TYPING_EXTENSIONS",
    "THIRDPARTY",
    "LOCALFOLDER",
    "FIRSTPARTY",
]
